optimized saving, npy at regular intervals

+++ PDBbind_data/similarity/pairwise_similarity_matrix/pairwise_similarity_tanimoto.py
@@ -7,7 +7,6 @@ import subprocess
 import logging
 from logging.handlers import RotatingFileHandler
 import warnings
-import h5py
 from scipy.spatial import KDTree
 from time import time
 from joblib import Parallel, delayed
@@ -79,6 +78,17 @@ def process_pair(id1, id2, mol1, mol2):
     return tanimoto
 
 
+def matrices_to_npy(similarity_matrix_tanimoto, PSM_tanimoto_file):
+        '''
+        Saving matrices to the disk as .npy files, overwriting existing matrices. 
+        Makes sure the np matrices are symmetric and the diagonal is correct
+        '''
+        similarity_matrix_tanimoto = np.maximum(similarity_matrix_tanimoto[:], similarity_matrix_tanimoto[:].T)
+        np.fill_diagonal(similarity_matrix_tanimoto, 1.0)
+        np.save(PSM_tanimoto_file, similarity_matrix_tanimoto)
+        print(f"Tanimoto similarity matrix saved to {PSM_tanimoto_file}", flush=True)
+
+
 
 def main(folder_path, save_as_json=False):
 
@@ -95,13 +105,10 @@ def main(folder_path, save_as_json=False):
                 json.dump(complexes, f)
             print("List of complexes saved to pairwise_similarity_complexes.json", flush=True)
 
-
-        # Initialize the HDF5 file and dataset to save the similarities
-        with h5py.File('pairwise_similarity_tanimoto.hdf5', 'a') as f:
-            if 'similarities' not in f:
-                dset = f.create_dataset("similarities", (num_complexes, num_complexes), dtype='float32', compression="gzip")
-            else:
-                dset = f['similarities']
+        # INITIALIZE PAIRWISE SIMILARITY MATRICES
+        PSM_tanimoto_file = "pairwise_similarity_matrix_tanimoto.npy"
+        print("Creating Tanimoto similarity matrix...", flush=True)
+        similarity_matrix_tanimoto = np.zeros((len(complexes), len(complexes)), dtype=np.float32)
 
 
         # Parse the SDF files and store the molecules in a dictionary
@@ -126,16 +133,12 @@ def main(folder_path, save_as_json=False):
             if os.path.exists(precomputed_scores):
                 
                 try:
-                    # Append tanimoto data to the h5py file
+                    # Append tanimoto data to the tanimoto matrix
                     with open(precomputed_scores, 'r') as f:
                         scores = json.load(f)
-                        scores = [scores[complex] for complex in complexes] # Convert values to list in the right order
-                    with h5py.File('pairwise_similarity_tanimoto.hdf5', 'a') as f:
-                        dset = f['similarities']
-                        dset[i, :] = scores
-                        dset[:, i] = scores
-
-
+                        scores = [scores[complex] for complex in complexes]
+                    similarity_matrix_tanimoto[i, :] = scores
+                    similarity_matrix_tanimoto[:, i] = scores
                     
                     print(f"--- Precomputed data found for {complexes[i]}, skipping...", flush=True)
                 
@@ -157,50 +160,35 @@ def main(folder_path, save_as_json=False):
                     results = Parallel(n_jobs=-1)(delayed(process_pair)(
                             complexes[i], complexes[j], parsed_molecules[complexes[i]], parsed_molecules[complexes[j]]) for j in to_compare)
 
-                    # APPEND THE TANIMOTO SCORES TO THE HDF5 FILE
-                    with h5py.File('pairwise_similarity_tanimoto.hdf5', 'a') as f:
-                        dset = f['similarities']
-                        for j, metrics in zip(to_compare, results):
-                            dset[i, j] = metrics
-                            dset[j, i] = metrics
+                    # APPEND THE DATA TO THE SIMILARITY MATRICES
+                    for j, metrics in zip(to_compare, results):
+                        similarity_matrix_tanimoto[i, j] = metrics
+                        similarity_matrix_tanimoto[j, i] = metrics
 
                 if save_as_json:
 
                     # Save the tanimoto scores to a JSON file
-                    with h5py.File('pairwise_similarity_tanimoto.hdf5', 'a') as f:
-                        dset = f['similarities']
+                    row = similarity_matrix_tanimoto[i, :].tolist()
+                    sim_data = {complexes[j]: row[j] for j in range(num_complexes)}
+                    if not os.path.exists(precomputed_scores):
+                        with open(precomputed_scores, 'w') as f:
+                            json.dump(sim_data, f, indent=4)
 
-                        line_data = np.array(dset[i, :], dtype=np.float32)
-                        line_data = line_data.tolist()
-                        sim_data = {complexes[j]: line_data[j] for j in range(num_complexes)}
-                    
-                    # Save the dictionary to a JSON file with indentation
-                    with open(precomputed_scores, 'w') as f:
-                        json.dump(sim_data, f, indent=4)
-                        
                     print(f"--- Saved the tanimoto scores for {complexes[i]} to {precomputed_scores}", flush=True)
 
                 toc = time()
                 print(f"--- Done: Time needed: {toc - tac:.2f} seconds. Total Time: {toc - tic}", flush=True)
+
+            # Save the similarity matrices to .npy files every 1000 complexes
+            if (i + 1) % 1000 == 0: 
+                matrices_to_npy(similarity_matrix_tanimoto, PSM_tanimoto_file)
             # --------------------------------------------------------------------------------------------------------------
 
 
-        # Make sure the HDF5 files are symmetric and the diagonal is correct
-        with h5py.File(f'pairwise_similarity_rmsd_ligand.hdf5', 'a') as f:
-            dset = f['similarities']
-            dset[:] = np.maximum(dset[:], dset[:].T)
+        # Save the final similarity matrices to .npy files
+        matrices_to_npy(similarity_matrix_tanimoto, PSM_tanimoto_file)
 
-            arr = dset[:]
-            np.fill_diagonal(arr, 0.0)
-            dset[:] = arr
 
-        with h5py.File(f'pairwise_similarity_tm_scores.hdf5', 'a') as f:
-            dset = f['similarities']
-            dset[:] = np.maximum(dset[:], dset[:].T)
-
-            arr = dset[:]
-            np.fill_diagonal(arr, 1.0)
-            dset[:] = arr
 
     except Exception as e:
         logger.error(f"Error in main function: {str(e)}")
